requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0

streamlit==1.32.0
requests==2.31.0
boto3==1.35.74